Make fuzz command in mill
